+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fix Drop component does not close on pointerleave in Firefox
+- Fix Switcher component sets active state on ignored item
 
 ## 3.3.3 (February 25, 2020)